fixed numeric type and test cases

+++ src/main/java/com/eppo/sdk/deserializer/EppoValueDeserializer.java
@@ -72,7 +72,7 @@ public class EppoValueDeserializer extends StdDeserializer<EppoValue> {
                 }
                 return EppoValue.valueOf(array);
             case NUMBER:
-                return EppoValue.valueOf(node.asLong());
+                return EppoValue.valueOf(node.asDouble());
             case STRING:
                 return EppoValue.valueOf(node.asText());
             case BOOLEAN:
